(BSR) fix(retrieval): semantic vector retrieval build  (#4264)
* added load method in textclient

* modifies load functions in clients for clarity

* add prediction example for semantic retrueval in readme

* fixed add_item_similarities function to work with textclient which does not have user_docs

* moved add_items_dot_similarity function to recoclient

* PR comments

* add kwargs

* add tensorflow macos deps to pyproject.toml
